refactor(discord): extract provider error presentation
